PSY-1017: remove dead app/nav.tsx; move live user-display helpers out of nav-utils
The nav audit's "both files dead" claim was stale for nav-utils:
UserMenu.tsx (PSY-1018) imports getUserInitials/getUserDisplayName.
Move those two helpers (verbatim, with their tests) to a colocated
components/layout/nav/userDisplay.ts, then delete app/nav.tsx (zero
importers), app/nav-utils.ts, and app/nav-utils.test.ts. The
navLinks/isExternal/NavLink exports and their tests die with the
legacy nav.
